add vinyl typing to the source code instead of as a d.ts file, otherwise, extensions that use bit-bin throw tsc errors
