+++ tsconfig.json
@@ -13,7 +13,6 @@
     "esModuleInterop": true,
     "resolveJsonModule": true,
     "sourceMap": true,
-    "typeRoots": ["./custom-types", "./node_modules/@types"],
     "outDir": "./dist"
   },
   "include": [
